requirements.txt
rx
kafka-python
requests
python-geohash